bug: support floats

+++ src/calc/eval.rs
@@ -45,7 +45,7 @@ mod tests {
     use super::*;
     use crate::calc::token::{Operator, Token};
 
-    fn n(x: u32) -> Token {
+    fn n(x: f32) -> Token {
         Token::Number(x)
     }
 
@@ -55,7 +55,7 @@ mod tests {
 
     #[test]
     fn eval_add() {
-        let input = vec![n(1), n(2), op(Operator::Add)];
+        let input = vec![n(1.), n(2.), op(Operator::Add)];
         let res = eval_postfix(input).unwrap();
         assert_eq!(res, 3.0);
     }
@@ -63,7 +63,7 @@ mod tests {
     #[test]
     fn eval_all_ops() {
         // (5 + 5) * 2 = 20 → postfix: 5 5 + 2 *
-        let input = vec![n(5), n(5), op(Operator::Add), n(2), op(Operator::Mul)];
+        let input = vec![n(5.), n(5.), op(Operator::Add), n(2.), op(Operator::Mul)];
         let res = eval_postfix(input).unwrap();
         assert_eq!(res, 20.0);
     }
@@ -71,21 +71,21 @@ mod tests {
     #[test]
     fn eval_sub_order() {
         // 10 - 5 = 5 → postfix: 10 5 -
-        let input = vec![n(10), n(5), op(Operator::Sub)];
+        let input = vec![n(10.), n(5.), op(Operator::Sub)];
         let res = eval_postfix(input).unwrap();
         assert_eq!(res, 5.0);
     }
 
     #[test]
     fn eval_div() {
-        let input = vec![n(10), n(5), op(Operator::Div)];
+        let input = vec![n(10.), n(5.), op(Operator::Div)];
         let res = eval_postfix(input).unwrap();
         assert_eq!(res, 2.0);
     }
 
     #[test]
     fn eval_div_by_zero() {
-        let input = vec![n(10), n(0), op(Operator::Div)];
+        let input = vec![n(10.), n(0.), op(Operator::Div)];
         let err = eval_postfix(input).unwrap_err();
         assert!(matches!(err, CalcError::DivisionByZero));
     }
@@ -100,14 +100,14 @@ mod tests {
     #[test]
     fn eval_partial_expression() {
         // 1 +
-        let input = vec![n(1), op(Operator::Add)];
+        let input = vec![n(1.), op(Operator::Add)];
         let err = eval_postfix(input).unwrap_err();
         assert!(matches!(err, CalcError::NotEnoughOperands));
     }
 
     #[test]
     fn eval_extra_operands_should_fail() {
-        let input = vec![n(1), n(2), n(3), op(Operator::Add)];
+        let input = vec![n(1.), n(2.), n(3.), op(Operator::Add)];
         let err = eval_postfix(input).unwrap_err();
         assert!(matches!(err, CalcError::NotEnoughOperands));
     }
